Update paper.tex
The velocity graph was based on the wrong equation for redshift to velocity. In fact, since the graph would lead to velocities greater than the speed of light, that suggests that a stretching of space model may be a better fit than a recessional velocity model. Either way, the graph was wrong and need to be revisited.